refactor: move NotAllowedError and ResourceNotFoundError to /core
