+++ src/core/errors/errors/not-allowed-error.ts
