+++ src/core/errors/errors/resource-not-found.ts
