+++ src/domain/forum/application/use-cases/comment-on-answer.ts
@@ -3,7 +3,7 @@ import { UniqueEntityId } from '@/core/entities/value-objects/unique-entity-id'
 import { AnswerComment } from '@/domain/forum/enterprise/entities/answer-comment'
 import { AnswersCommentsRepository } from '@/domain/forum/application/repositories/answers-comments.repository'
 import { Either, left, right } from '@/core/either'
-import { ResourceNotFoundError } from '@/domain/forum/application/use-cases/errors/resource-not-found'
+import { ResourceNotFoundError } from '@/core/errors/errors/resource-not-found'
 
 interface CommentOnAnswerRequest {
   authorId: string
