+++ src/domain/forum/application/use-cases/delete-answer-comment.spec.ts
@@ -2,7 +2,7 @@ import { InMemoryAnswerCommentsRepository } from '@root/test/repositories/in-mem
 import { DeleteAnswerCommentUseCase } from '@/domain/forum/application/use-cases/delete-answer-comment'
 import { makeAnswerComment } from '@root/test/factories/make-answer-comment'
 import { UniqueEntityId } from '@/core/entities/value-objects/unique-entity-id'
-import { NotAllowedError } from '@/domain/forum/application/use-cases/errors/not-allowed-error'
+import { NotAllowedError } from '@/core/errors/errors/not-allowed-error'
 
 let inMemoryAnswerCommentsRepository: InMemoryAnswerCommentsRepository
 let sut: DeleteAnswerCommentUseCase
