+++ src/domain/forum/application/use-cases/delete-answer.ts
@@ -1,7 +1,7 @@
 import { AnswersRepository } from '@/domain/forum/application/repositories/answers.repository'
 import { Either, left, right } from '@/core/either'
-import { ResourceNotFoundError } from '@/domain/forum/application/use-cases/errors/resource-not-found'
-import { NotAllowedError } from '@/domain/forum/application/use-cases/errors/not-allowed-error'
+import { ResourceNotFoundError } from '@/core/errors/errors/resource-not-found'
+import { NotAllowedError } from '@/core/errors/errors/not-allowed-error'
 
 interface DeleteAnswerRequest {
   authorId: string
