+++ src/domain/forum/application/use-cases/edit-answer.ts
@@ -1,7 +1,7 @@
 import { AnswersRepository } from '@/domain/forum/application/repositories/answers.repository'
 import { Answer } from '@/domain/forum/enterprise/entities/answer'
-import { ResourceNotFoundError } from '@/domain/forum/application/use-cases/errors/resource-not-found'
-import { NotAllowedError } from '@/domain/forum/application/use-cases/errors/not-allowed-error'
+import { ResourceNotFoundError } from '@/core/errors/errors/resource-not-found'
+import { NotAllowedError } from '@/core/errors/errors/not-allowed-error'
 import { Either, left, right } from '@/core/either'
 import { UniqueEntityId } from '@/core/entities/value-objects/unique-entity-id'
 import { AnswerAttachment } from '@/domain/forum/enterprise/entities/answer-attachment'
