+++ src/domain/forum/application/use-cases/edit-question.spec.ts
@@ -1,5 +1,5 @@
 import { EditQuestionUseCase } from './edit-question'
-import { NotAllowedError } from '@/domain/forum/application/use-cases/errors/not-allowed-error'
+import { NotAllowedError } from '@/core/errors/errors/not-allowed-error'
 import { InMemoryQuestionsRepository } from '@root/test/repositories/in-memory-questions.repository'
 import { InMemoryQuestionAttachmentsRepository } from '@root/test/repositories/in-memory-question-attachments.repository'
 import { makeQuestion } from '@root/test/factories/make-questions'
